# Conflicts:
#	src/main/java/com/example/joinair/controller/AdminController.java
#	src/main/java/com/example/joinair/controller/NoticeController.java
#	src/main/java/com/example/joinair/service/AdminService.java
#	src/main/resources/application.properties
#	src/main/resources/static/js/searchProductad.js
#	src/main/resources/templates/SalesReport.html
#	src/main/resources/templates/about.html
#	src/main/resources/templates/adminEditUser.html
#	src/main/resources/templates/adminEditUserList.html
#	src/main/resources/templates/cart/index.html
#	src/main/resources/templates/checkpaymentinformation.html
#	src/main/resources/templates/contact.html
#	src/main/resources/templates/index.html
#	src/main/resources/templates/index2.html
#	src/main/resources/templates/membership.html
#	src/main/resources/templates/noticelist2.html
#	src/main/resources/templates/noticeview.html
#	src/main/resources/templates/ordercomplete.html
#	src/main/resources/templates/productadlist.html
#	src/main/resources/templates/productbuylist.html
#	src/main/resources/templates/productbuylist2.html

+++ src/main/java/com/example/joinair/dto/ORDER_DETAIL.java
@@ -12,13 +12,14 @@ public class ORDER_DETAIL {
     private String OD_Pro_Name; // 주문한 상품명    ( DB컬럼 생성 필요함)
     private int OD_Pro_Price; // 주문한 상품명의 단가  (DB컬럼 생성 필요함)
     private int OD_Qua; // 수량             (session에 담긴 Quantity)
-    private double OD_Totalweight; // 총 무게     (session에 담긴 weight)
-    private double OD_Total_Price; // 총 소계금액  (session에 담긴 Subtotal)
+    private int OD_Totalweight; // 총 무게     (session에 담긴 weight)
+    private int OD_Total_Price; // 총 소계금액  (session에 담긴 Subtotal)
 
 
 
 
-    private Long Orders_Num; // 주문번호    (x)
+
+    private int Orders_Num; // 주문번호    (x)
     private String OD_Location; // 위치             (x)
     private Date OD_Reqtime; // 요청배달시간         (X)
     private String OD_Request; // 배송요청사항   (X)
